+++ app/planetgames/components/ProductCarousel.jsx
@@ -37,8 +37,8 @@ const ProductCarousel = ({ title, products }) => {
   }, []);
 
   return (
-    <div className="mt-16">
-      <h2 className="text-2xl font-bold text-gray-900 mb-6">{title}</h2>
+    <div className="mt-16 p-2">
+      <h2 className="text-3xl font-bold text-gray-900 mb-6">{title}</h2>
 
       {/* المنتجات */}
       <div
