+++ app/planetgames/details_of_product/[id_p]/layout.js
